+++ docs/CHANGELOG.md
@@ -5,13 +5,24 @@
 ## 📦 Bookshelf - 2.2.0
 
 ### 🎉 Addition
-
 - ([187](https://github.com/Gunivers/Bookshelf/pull/222)) Add randomized fill function `#bs.block:fill_random`.
 - ([1](https://github.com/Gunivers/Bookshelf/issues/1)) ([220](https://github.com/Gunivers/Bookshelf/issues/220)) Add `bs.random` module for random distributions and noise generation.
 
-### ⚡️ Improvements
+---
 
+## 📦 Bookshelf - 2.1.1
+
+### 🔁 Tweaks
+- ([238](https://github.com/Gunivers/Bookshelf/pull/238)) Load tags now load modules in isolation.
+- ([237](https://github.com/Gunivers/Bookshelf/pull/237)) Enhanced speed of `bs.block` module.
 - ([218](https://github.com/Gunivers/Bookshelf/pull/218)) Enhanced speed of factorial function in `bs.math`.
+
+### 🐛 Fixes
+- ([236](https://github.com/Gunivers/Bookshelf/issues/236)) Fix `bs.block`, `bs.id` and `bs.math` load functions.
+- ([235](https://github.com/Gunivers/Bookshelf/issues/235)) Fix `bs.block` inputs that required extra escaping.
+
+### 📄 Documentation
+- ([215](https://github.com/Gunivers/Bookshelf/issues/215)) There is now documentation about the documentation (kind of metadocumentation?).
 
 ---